backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0
anthropic==0.42.0
pydantic==2.10.3
pydantic-settings==2.7.0
python-multipart==0.0.19
python-docx==1.1.2
openpyxl==3.1.5
pymupdf==1.25.1
aiofiles==24.1.0
httpx==0.28.1
python-dotenv==1.0.1
fpdf2==2.8.2
sse-starlette==2.1.3
